Prevent recursive loop when NXDOMAIN
When the answer is NXDOMAIN, GoDNS tries to reach upstream. Upstream was configured with the local dnsmasq which forwards the request back to GoDNS as it's a domain name in *.scalingo.dbs.test. And here is a infinite loop. I updated the upstream server to be Cloudflare DNS which will definitely return NXDOMAIN for such resquests